[REFACTOR] Moved view and card presenters to correct package. Follows `leanback-showcase` apps package pattern.
