+++ tv/src/main/java/info/hossainkhan/dailynewsheadlines/cards/presenters/ImageCardViewPresenter.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package info.hossainkhan.dailynewsheadlines;
+package info.hossainkhan.dailynewsheadlines.cards.presenters;
 
 import android.content.Context;
 import android.support.v17.leanback.widget.ImageCardView;
@@ -30,6 +30,7 @@ import android.view.ContextThemeWrapper;
 
 import com.squareup.picasso.Picasso;
 
+import info.hossainkhan.dailynewsheadlines.R;
 import io.swagger.client.model.Article;
 
 /**
