+++ tv/src/main/java/info/hossainkhan/dailynewsheadlines/cards/presenters/TextCardPresenter.java
@@ -22,10 +22,12 @@
  * SOFTWARE.
  */
 
-package info.hossainkhan.dailynewsheadlines;
+package info.hossainkhan.dailynewsheadlines.cards.presenters;
 
 import android.content.Context;
 
+import info.hossainkhan.dailynewsheadlines.AbstractCardPresenter;
+import info.hossainkhan.dailynewsheadlines.cards.TextCardView;
 import io.swagger.client.model.Article;
 
 /**
